[Custom images] Enable build to happen on instance in shared VPC
The current code enforces the use of the default project network
when generating the VM for the build.

This is not suitable when the instance VM for the build needs to
be placed in a shared VPC, where only the subnetwork argument
is required.

This patch allows the support for this feature by preventing the
default network to be set when only a subnetwork argument is provided.